# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dummy Backend API",
    "description": "API documentation for the Dummy Backend system - managing HR, Payroll, Operations, Inventory, and Fleet Management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
...
  "paths": {
...
    "/clean/rentals": {
      "get": {
        "tags": ["Rentals"],
        "summary": "Get all rentals or a specific rental",
        "description": "Returns rental assignments with bus and employee details",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Rental Assignment ID",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by rental status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "confirmed", "in_progress", "completed", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/RentalAssignment" }
                },
                "example": [
                  {
                    "assignment_id": "RA-ppt19z6wl6tapo4bntaqurm4",
                    "bus_id": "BUS021",
                    "bus_plate_number": "PLATE-0021",
                    "bus_type": "Aircon",
                    "body_number": "BODY-0021",
                    "rental_status": "completed",
                    "rental_details": {
                      "rental_id": "RENT-abc123",
                      "rental_type": "Corporate Event",
                      "rental_date": "2026-01-15T00:00:00.000Z",
                      "return_date": "2026-01-16T00:00:00.000Z",
                      "pickup_location": "Sapang Palay Terminal",
                      "dropoff_location": "MOA Complex",
                      "total_amount": 8500,
                      "notes": null
                    },
                    "employees": [
                      {
                        "employeeId": "clxyz789",
                        "employeeNumber": "EMP-2021-OPS-D001",
                        "firstName": "Mario",
                        "lastName": "Perez",
                        "role": "Driver"
                      }
                    ]
                  }
                ]
              }
            }
          },
          "404": { "description": "Rental not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Rentals"],
        "summary": "Create a new rental",
        "description": "Create a new bus rental with optional bus and employee assignments",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RentalCreate" },
              "example": {
                "rental_date": "2026-02-01",
                "return_date": "2026-02-02",
                "rental_type": "Wedding",
                "pickup_location": "SM City San Jose",
                "dropoff_location": "Tagaytay City",
                "total_amount": 12000,
                "bus_id": "BUS021",
                "employee_ids": [
                  { "id": "employee-cuid-1", "role": "Driver" },
                  { "id": "employee-cuid-2", "role": "Conductor" }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rental created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assignment_id": { "type": "string" },
                    "rental_id": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "500": { "description": "Internal server error" }
        }
      },
      "patch": {
        "tags": ["Rentals"],
        "summary": "Update a rental",
        "description": "Update rental status, details, or bus assignment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assignment_id"],
                "properties": {
                  "assignment_id": { "type": "string" },
                  "status": { "type": "string" },
                  "rental_type": { "type": "string" },
                  "notes": { "type": "string" },
                  "bus_id": { "type": "string" }
                }
              },
              "example": {
                "assignment_id": "RA-ppt19z6wl6tapo4bntaqurm4",
                "status": "completed",
                "notes": "Trip completed successfully"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Rental updated successfully" },
          "400": { "description": "Invalid request" },
          "404": { "description": "Rental not found" }
        }
      }
    },
...
    "/status": {
      "get": {
        "tags": ["System"],
        "summary": "Get system status",
        "description": "Returns database connection status and table counts",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SystemStatus" }
              }
            }
          },
          "500": { "description": "Database connection failed" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "RentalAssignment": {
        "type": "object",
        "properties": {
          "assignment_id": { "type": "string" },
          "bus_id": { "type": "string", "nullable": true },
          "bus_plate_number": { "type": "string" },
          "bus_type": { "type": "string" },
          "body_number": { "type": "string" },
          "rental_status": { "type": "string", "enum": ["pending", "confirmed", "in_progress", "completed", "cancelled"] },
          "rental_details": { "$ref": "#/components/schemas/RentalDetails" },
          "employees": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RentalEmployee" }
          }
        }
      },
      "RentalDetails": {
        "type": "object",
        "properties": {
          "rental_id": { "type": "string" },
          "rental_type": { "type": "string" },
          "rental_date": { "type": "string", "format": "date-time" },
          "return_date": { "type": "string", "format": "date-time", "nullable": true },
          "pickup_location": { "type": "string" },
          "dropoff_location": { "type": "string" },
          "total_amount": { "type": "number" },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "RentalEmployee": {
        "type": "object",
        "properties": {
          "employeeId": { "type": "string" },
          "employeeNumber": { "type": "string" },
          "firstName": { "type": "string" },
          "middleName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "position": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "RentalCreate": {
        "type": "object",
        "required": ["rental_date"],
        "properties": {
          "rental_date": { "type": "string", "format": "date" },
          "return_date": { "type": "string", "format": "date" },
          "rental_type": { "type": "string" },
          "pickup_location": { "type": "string" },
          "dropoff_location": { "type": "string" },
          "total_amount": { "type": "number" },
          "notes": { "type": "string" },
          "bus_id": { "type": "string" },
          "employee_ids": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "role": { "type": "string" }
              }
            }
          }
        }
      },
...
      "SystemStatus": {
        "type": "object",
        "properties": {
          "database_status": { "type": "string" },
          "tables": {
            "type": "object",
            "properties": {
              "employees": { "type": "integer" },
              "departments": { "type": "integer" },
              "routes": { "type": "integer" },
              "buses": { "type": "integer" },
              "bus_assignments": { "type": "integer" }
            }
          },
          "endpoints": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}